Add a silent flag
The silent flag can be added to any condition. If it is set and the spec
fails because of this condition, the info is passed to the caller. Then
the caller can decide what to do with the info. It is useful to silently
"ignore" rules so the challenges won't get declined, for example to
leave a cheater in the dark.